% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memplot.R
\name{mem_plot}
\alias{mem_plot}
\title{Marginal effects at the mean plots}
\usage{
mem_plot(model, variable, plt_range, collector = c("typical", "average"))
}
\arguments{
\item{model}{a model object with a data slot and a predict() method}

\item{variable}{a character of the variable of interest}

\item{plt_range}{optional vector of length 2 indicating the range to plot
over for the x - axis}

\item{collector}{"typical" or "average" or an abbreviation thereof. See
details}
}
\value{
`ggplot` object with the MEM plot
}
\description{
The better way of interpreting coefficients in a model: take a typical
sample, vary the variable of interest, and see how that changes the
prediction. Works for categorical and continuous predictors.
}
\details{
The `typical` collector calculates the median values for numeric,
and the most common class for categorical predictors. The `average` collector
returns the mean for numeric variables and the reference category for
categorical variables.
}
\examples{
mod <- glm(vs ~ drat + qsec, family = "binomial", data = mtcars)
mem_plot(mod, "qsec")
mem_plot(mod, "drat")
}
