% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memplot.R
\name{typical_case}
\alias{typical_case}
\title{Find typical case}
\usage{
typical_case(data, collector = c("typical", "average"))
}
\arguments{
\item{data}{dataset to find the typical values in}

\item{collector}{"average" or "typical" (default)}
}
\value{
a named list of typical values for each variable
}
\description{
Function to find the typical value for each variable as used in mem_plot
}
